Move things to the right place
